requirements.txt
autopep8==1.4.3
dj-database-url==0.5.0
Django==2.2
django-rest-framework==0.1.0
djangorestframework==3.9.2
entrypoints==0.3
flake8==3.7.7
gunicorn==19.9.0
mccabe==0.6.1
psycopg2==2.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.1
sqlparse==0.3.0
